Explicitly deallocate memory in fluxes and optics
Add routines to explicitly free memory in the flux and optics objects. I
was running into an out of memory error in ne30 simulations, which I
think was due to how I was handling the ty_fluxes class. The class
declares member data as pointers, but I was using them as allocatables
without explicitly deallocting the memory when I was done with them. I
think this resulted in inaccessible memory after leaving the routine,
since the "allocate" would have allocated memory for the pointer data,
but with no way to access that memory after the routine exited and the
pointers were nullified.